Restructure canvasser UI with separate page for households
